Create frontend for cooking app
Added index.html and index.js for the cooking app frontend. The frontend features form inputs for recipe attributes such as name, day, and type. It also includes two lists: one for available ingredients and another for chosen ingredients, along with a cook button to add a new recipe. Further, it fetches data from the server to display saved recipes.